+++ src/Core/WorkSpace/src/Components/NoDataViewer/NoDataStyles.ts
