+++ src/Core/WorkSpace/src/Components/WorkSpace/WorkSpaceViewStyles.ts
