+++ src/Core/WorkSpace/src/Components/WorkSpaceTabs/WorkSpaceTabsStyles.ts
