+++ src/Core/WorkSpace/src/Context/DataProvider.tsx
@@ -1,7 +1,7 @@
 import { AnalyticsOptions } from '@equinor/Diagrams';
 import { createContext, useCallback, useContext, useEffect, useReducer } from 'react';
 import { ActionType, createCustomAction, getType } from 'typesafe-actions';
-import { useWorkSpaceKey } from '../Components/DefaultDataView/Hooks/useDataViewerKey';
+import { useWorkSpaceKey } from '../Components/DefaultView/Hooks/useDataViewerKey';
 import {
     DataViewSideSheetOptions,
     FilterOptions,
